+++ sycl/include/sycl/detail/range_rounding.hpp
@@ -12,13 +12,13 @@
 #include <sycl/detail/export.hpp>
 #include <sycl/detail/helpers.hpp>
 #include <sycl/detail/iostream_proxy.hpp>
-#include <sycl/device.hpp>
 #include <sycl/ext/oneapi/kernel_properties.hpp>
 #include <sycl/id.hpp>
 #include <sycl/item.hpp>
 #include <sycl/kernel_handler.hpp>
 #include <sycl/range.hpp>
 
+#include <limits>
 #include <tuple>
 #include <type_traits>
 
@@ -26,6 +26,7 @@
 
 namespace sycl {
 inline namespace _V1 {
+class device;
 
 namespace detail {
