+++ sycl/test-e2e/HierPar/Inputs/hier_par_wgscope_impl.hpp
@@ -11,6 +11,7 @@
 #include <iostream>
 #include <memory>
 #include <sycl/detail/core.hpp>
+#include <sycl/h_item.hpp>
 
 using namespace sycl;
