+++ sycl/test-e2e/InorderQueue/in_order_kernels.cpp
@@ -14,6 +14,7 @@
 #include <iostream>
 
 #include <sycl/detail/core.hpp>
+#include <sycl/h_item.hpp>
 
 #include <sycl/properties/all_properties.hpp>
 #include <sycl/usm.hpp>
