+++ sycl/test-e2e/Properties/cache_config.cpp
@@ -22,6 +22,7 @@
 #include <numeric>
 #include <sycl/detail/core.hpp>
 #include <sycl/ext/oneapi/experimental/enqueue_functions.hpp>
+#include <sycl/h_item.hpp>
 #include <sycl/ext/oneapi/free_function_queries.hpp>
 #include <sycl/kernel_bundle.hpp>
 #include <sycl/properties/all_properties.hpp>
