+++ sycl/test/include_deps/sycl_detail_core.hpp.cpp
@@ -14,10 +14,11 @@
 // CHECK-NEXT: id.hpp
 // CHECK-NEXT: detail/array.hpp
 // CHECK-NEXT: range.hpp
+// CHECK-NEXT: detail/code_location.hpp
+// CHECK-NEXT: detail/export.hpp
 // CHECK-NEXT: detail/common.hpp
 // CHECK-NEXT: detail/assert.hpp
 // CHECK-NEXT: __spirv/spirv_vars.hpp
-// CHECK-NEXT: detail/export.hpp
 // CHECK-NEXT: detail/nd_loop.hpp
 // CHECK-NEXT: detail/fwd/buffer.hpp
 // CHECK-NEXT: detail/generic_type_traits.hpp
@@ -89,18 +90,10 @@
 // CHECK-NEXT: detail/backend_types_io.hpp
 // CHECK-NEXT: queue.hpp
 // CHECK-NEXT: detail/cg_types.hpp
-// CHECK-NEXT: detail/host_profiling_info.hpp
-// CHECK-NEXT: detail/item_base.hpp
-// CHECK-NEXT: detail/kernel_desc.hpp
-// CHECK-NEXT: group.hpp
-// CHECK-NEXT: detail/async_work_group_copy_ptr.hpp
-// CHECK-NEXT: device_event.hpp
-// CHECK-NEXT: h_item.hpp
-// CHECK-NEXT: item.hpp
 // CHECK-NEXT: kernel_handler.hpp
-// CHECK-NEXT: nd_item.hpp
-// CHECK-NEXT: nd_range.hpp
+// CHECK-NEXT: detail/kernel_desc.hpp
 // CHECK-NEXT: detail/nd_range_view.hpp
+// CHECK-NEXT: nd_range.hpp
 // CHECK-NEXT: detail/optional.hpp
 // CHECK-NEXT: detail/range_rounding.hpp
 // CHECK-NEXT: device.hpp
@@ -115,6 +108,13 @@
 // CHECK-NEXT: ext/oneapi/properties/property.hpp
 // CHECK-NEXT: ext/oneapi/properties.hpp
 // CHECK-NEXT: ext/oneapi/properties/property_utils.hpp
+// CHECK-NEXT: item.hpp
+// CHECK-NEXT: detail/item_base.hpp
+// CHECK-NEXT: device.hpp
+// CHECK-NEXT: detail/string_view.hpp
+// CHECK-NEXT: detail/util.hpp
+// CHECK-NEXT: device_selector.hpp
+// CHECK-NEXT: kernel_bundle_enums.hpp
 // CHECK-NEXT: event.hpp
 // CHECK-NEXT: exception_list.hpp
 // CHECK-NEXT: ext/oneapi/device_global/device_global.hpp
@@ -148,14 +148,16 @@
 // CHECK-NEXT: ext/oneapi/work_group_scratch_memory.hpp
 // CHECK-NEXT: detail/sycl_local_mem_builtins.hpp
 // CHECK-NEXT: detail/reduction_forward.hpp
-// CHECK-NEXT: detail/ur.hpp
-// CHECK-NEXT: ur_api_funcs.def
 // CHECK-NEXT: ext/oneapi/bindless_images_interop.hpp
 // CHECK-NEXT: ext/oneapi/interop_common.hpp
 // CHECK-NEXT: ext/oneapi/bindless_images_mem_handle.hpp
 // CHECK-NEXT: ext/oneapi/experimental/free_function_traits.hpp
 // CHECK-NEXT: ext/oneapi/experimental/raw_kernel_arg.hpp
+// CHECK-NEXT: group.hpp
+// CHECK-NEXT: detail/async_work_group_copy_ptr.hpp
+// CHECK-NEXT: device_event.hpp
 // CHECK-NEXT: kernel.hpp
+// CHECK-NEXT: nd_item.hpp
 // CHECK-NEXT: sampler.hpp
 // CHECK-NEXT: sycl_span.hpp
 // CHECK-NEXT: feature_test.hpp
