+++ sycl/test/include_deps/sycl_khr_split_headers_kernel_bundle.hpp.cpp
@@ -76,6 +76,7 @@
 // CHECK-NEXT: accessor.hpp
 // CHECK-NEXT: detail/accessor_iterator.hpp
 // CHECK-NEXT: detail/fwd/accessor.hpp
+// CHECK-NEXT: detail/code_location.hpp
 // CHECK-NEXT: detail/common.hpp
 // CHECK-NEXT: detail/assert.hpp
 // CHECK-NEXT: detail/nd_loop.hpp
@@ -99,18 +100,9 @@
 // CHECK-NEXT: detail/compile_time_kernel_info.hpp
 // CHECK-NEXT: detail/kernel_launch_helper.hpp
 // CHECK-NEXT: detail/cg_types.hpp
+// CHECK-NEXT: kernel_handler.hpp
 // CHECK-NEXT: detail/helpers.hpp
 // CHECK-NEXT: memory_enums.hpp
-// CHECK-NEXT: detail/host_profiling_info.hpp
-// CHECK-NEXT: detail/item_base.hpp
-// CHECK-NEXT: group.hpp
-// CHECK-NEXT: detail/async_work_group_copy_ptr.hpp
-// CHECK-NEXT: device_event.hpp
-// CHECK-NEXT: h_item.hpp
-// CHECK-NEXT: item.hpp
-// CHECK-NEXT: kernel_handler.hpp
-// CHECK-NEXT: nd_item.hpp
-// CHECK-NEXT: nd_range.hpp
 // CHECK-NEXT: detail/is_device_copyable.hpp
 // CHECK-NEXT: ext/intel/experimental/fp_control_kernel_properties.hpp
 // CHECK-NEXT: ext/oneapi/properties/property.hpp
@@ -119,6 +111,26 @@
 // CHECK-NEXT: ext/oneapi/experimental/cluster_group_prop.hpp
 // CHECK-NEXT: ext/oneapi/properties.hpp
 // CHECK-NEXT: ext/oneapi/properties/property_utils.hpp
+// CHECK-NEXT: ext/oneapi/experimental/use_root_sync_prop.hpp
+// CHECK-NEXT: ext/oneapi/experimental/virtual_functions.hpp
+// CHECK-NEXT: ext/oneapi/free_function_kernel_properties.hpp
+// CHECK-NEXT: ext/oneapi/kernel_properties.hpp
+// CHECK-NEXT: ext/oneapi/work_group_scratch_memory.hpp
+// CHECK-NEXT: detail/sycl_local_mem_builtins.hpp
+// CHECK-NEXT: detail/nd_range_view.hpp
+// CHECK-NEXT: nd_range.hpp
+// CHECK-NEXT: detail/range_rounding.hpp
+// CHECK-NEXT: detail/iostream_proxy.hpp
+// CHECK-NEXT: item.hpp
+// CHECK-NEXT: detail/item_base.hpp
+// CHECK-NEXT: detail/reduction_forward.hpp
+// CHECK-NEXT: event.hpp
+// CHECK-NEXT: ext/oneapi/bindless_images_interop.hpp
+// CHECK-NEXT: ext/oneapi/interop_common.hpp
+// CHECK-NEXT: ext/oneapi/bindless_images_mem_handle.hpp
+// CHECK-NEXT: ext/oneapi/device_global/device_global.hpp
+// CHECK-NEXT: ext/oneapi/device_global/properties.hpp
+// CHECK-NEXT: ext/oneapi/experimental/free_function_traits.hpp
 // CHECK-NEXT: ext/oneapi/experimental/graph.hpp
 // CHECK-NEXT: ext/oneapi/experimental/graph/command_graph.hpp
 // CHECK-NEXT: ext/oneapi/experimental/graph/common.hpp
@@ -130,24 +142,11 @@
 // CHECK-NEXT: ext/oneapi/experimental/graph/modifiable_graph.hpp
 // CHECK-NEXT: ext/oneapi/experimental/graph/dynamic.hpp
 // CHECK-NEXT: ext/oneapi/experimental/work_group_memory.hpp
-// CHECK-NEXT: ext/oneapi/experimental/use_root_sync_prop.hpp
-// CHECK-NEXT: ext/oneapi/experimental/virtual_functions.hpp
-// CHECK-NEXT: ext/oneapi/free_function_kernel_properties.hpp
-// CHECK-NEXT: ext/oneapi/kernel_properties.hpp
-// CHECK-NEXT: ext/oneapi/work_group_scratch_memory.hpp
-// CHECK-NEXT: detail/sycl_local_mem_builtins.hpp
-// CHECK-NEXT: detail/nd_range_view.hpp
-// CHECK-NEXT: detail/range_rounding.hpp
-// CHECK-NEXT: detail/iostream_proxy.hpp
-// CHECK-NEXT: detail/reduction_forward.hpp
-// CHECK-NEXT: event.hpp
-// CHECK-NEXT: ext/oneapi/bindless_images_interop.hpp
-// CHECK-NEXT: ext/oneapi/interop_common.hpp
-// CHECK-NEXT: ext/oneapi/bindless_images_mem_handle.hpp
-// CHECK-NEXT: ext/oneapi/device_global/device_global.hpp
-// CHECK-NEXT: ext/oneapi/device_global/properties.hpp
-// CHECK-NEXT: ext/oneapi/experimental/free_function_traits.hpp
 // CHECK-NEXT: ext/oneapi/experimental/raw_kernel_arg.hpp
+// CHECK-NEXT: group.hpp
+// CHECK-NEXT: detail/async_work_group_copy_ptr.hpp
+// CHECK-NEXT: device_event.hpp
+// CHECK-NEXT: nd_item.hpp
 // CHECK-NEXT: sampler.hpp
 // CHECK-NEXT: sycl_span.hpp
 // CHECK-NEXT: specialization_id.hpp
