+++ sycl/test/include_deps/sycl_khr_split_headers_stream.hpp.cpp
@@ -17,10 +17,11 @@
 // CHECK-NEXT: id.hpp
 // CHECK-NEXT: detail/array.hpp
 // CHECK-NEXT: range.hpp
+// CHECK-NEXT: detail/code_location.hpp
+// CHECK-NEXT: detail/export.hpp
 // CHECK-NEXT: detail/common.hpp
 // CHECK-NEXT: detail/assert.hpp
 // CHECK-NEXT: __spirv/spirv_vars.hpp
-// CHECK-NEXT: detail/export.hpp
 // CHECK-NEXT: detail/nd_loop.hpp
 // CHECK-NEXT: detail/fwd/buffer.hpp
 // CHECK-NEXT: detail/generic_type_traits.hpp
@@ -96,10 +97,7 @@
 // CHECK-NEXT: detail/string_view.hpp
 // CHECK-NEXT: detail/kernel_launch_helper.hpp
 // CHECK-NEXT: detail/cg_types.hpp
-// CHECK-NEXT: detail/host_profiling_info.hpp
 // CHECK-NEXT: kernel_handler.hpp
-// CHECK-NEXT: nd_item.hpp
-// CHECK-NEXT: nd_range.hpp
 // CHECK-NEXT: detail/is_device_copyable.hpp
 // CHECK-NEXT: ext/intel/experimental/fp_control_kernel_properties.hpp
 // CHECK-NEXT: ext/oneapi/properties/property.hpp
@@ -108,17 +106,6 @@
 // CHECK-NEXT: ext/oneapi/experimental/cluster_group_prop.hpp
 // CHECK-NEXT: ext/oneapi/properties.hpp
 // CHECK-NEXT: ext/oneapi/properties/property_utils.hpp
-// CHECK-NEXT: ext/oneapi/experimental/graph.hpp
-// CHECK-NEXT: ext/oneapi/experimental/graph/command_graph.hpp
-// CHECK-NEXT: ext/oneapi/experimental/graph/common.hpp
-// CHECK-NEXT: ext/oneapi/experimental/graph/executable_graph.hpp
-// CHECK-NEXT: ext/oneapi/experimental/graph/node.hpp
-// CHECK-NEXT: ext/oneapi/experimental/detail/properties/graph_properties.hpp
-// CHECK-NEXT: ext/oneapi/experimental/detail/properties/graph_properties.def
-// CHECK-NEXT: ext/oneapi/experimental/detail/properties/node_properties.def
-// CHECK-NEXT: ext/oneapi/experimental/graph/modifiable_graph.hpp
-// CHECK-NEXT: ext/oneapi/experimental/graph/dynamic.hpp
-// CHECK-NEXT: ext/oneapi/experimental/work_group_memory.hpp
 // CHECK-NEXT: ext/oneapi/experimental/use_root_sync_prop.hpp
 // CHECK-NEXT: ext/oneapi/experimental/virtual_functions.hpp
 // CHECK-NEXT: ext/oneapi/free_function_kernel_properties.hpp
@@ -129,8 +116,10 @@
 // CHECK-NEXT: ext/oneapi/work_group_scratch_memory.hpp
 // CHECK-NEXT: detail/sycl_local_mem_builtins.hpp
 // CHECK-NEXT: detail/nd_range_view.hpp
+// CHECK-NEXT: nd_range.hpp
 // CHECK-NEXT: detail/range_rounding.hpp
 // CHECK-NEXT: detail/iostream_proxy.hpp
+// CHECK-NEXT: detail/reduction_forward.hpp
 // CHECK-NEXT: device.hpp
 // CHECK-NEXT: backend_types.hpp
 // CHECK-NEXT: detail/info_desc_helpers.hpp
@@ -152,10 +141,6 @@
 // CHECK-NEXT: detail/abi_neutral.hpp
 // CHECK-NEXT: device_selector.hpp
 // CHECK-NEXT: kernel_bundle_enums.hpp
-// CHECK-NEXT: detail/reduction_forward.hpp
-// CHECK-NEXT: detail/ur.hpp
-// CHECK-NEXT: detail/os_util.hpp
-// CHECK-NEXT: ur_api_funcs.def
 // CHECK-NEXT: event.hpp
 // CHECK-NEXT: ext/oneapi/bindless_images_interop.hpp
 // CHECK-NEXT: ext/oneapi/interop_common.hpp
@@ -163,8 +148,20 @@
 // CHECK-NEXT: ext/oneapi/device_global/device_global.hpp
 // CHECK-NEXT: ext/oneapi/device_global/properties.hpp
 // CHECK-NEXT: ext/oneapi/experimental/free_function_traits.hpp
+// CHECK-NEXT: ext/oneapi/experimental/graph.hpp
+// CHECK-NEXT: ext/oneapi/experimental/graph/command_graph.hpp
+// CHECK-NEXT: ext/oneapi/experimental/graph/common.hpp
+// CHECK-NEXT: ext/oneapi/experimental/graph/executable_graph.hpp
+// CHECK-NEXT: ext/oneapi/experimental/graph/node.hpp
+// CHECK-NEXT: ext/oneapi/experimental/detail/properties/graph_properties.hpp
+// CHECK-NEXT: ext/oneapi/experimental/detail/properties/graph_properties.def
+// CHECK-NEXT: ext/oneapi/experimental/detail/properties/node_properties.def
+// CHECK-NEXT: ext/oneapi/experimental/graph/modifiable_graph.hpp
+// CHECK-NEXT: ext/oneapi/experimental/graph/dynamic.hpp
+// CHECK-NEXT: ext/oneapi/experimental/work_group_memory.hpp
 // CHECK-NEXT: ext/oneapi/experimental/raw_kernel_arg.hpp
 // CHECK-NEXT: kernel.hpp
+// CHECK-NEXT: nd_item.hpp
 // CHECK-NEXT: sampler.hpp
 // CHECK-NEXT: usm/usm_enums.hpp
 // CHECK-EMPTY:
