+++ sycl/test/include_deps/sycl_khr_split_headers_usm.hpp.cpp
@@ -91,6 +91,7 @@
 // CHECK-NEXT: queue.hpp
 // CHECK-NEXT: accessor.hpp
 // CHECK-NEXT: detail/accessor_iterator.hpp
+// CHECK-NEXT: detail/code_location.hpp
 // CHECK-NEXT: detail/fwd/buffer.hpp
 // CHECK-NEXT: detail/handler_proxy.hpp
 // CHECK-NEXT: ext/oneapi/accessor_property_list.hpp
@@ -114,18 +115,10 @@
 // CHECK-NEXT: usm/usm_enums.hpp
 // CHECK-NEXT: properties/buffer_properties.def
 // CHECK-NEXT: detail/cg_types.hpp
-// CHECK-NEXT: detail/host_profiling_info.hpp
-// CHECK-NEXT: detail/item_base.hpp
-// CHECK-NEXT: detail/kernel_desc.hpp
-// CHECK-NEXT: group.hpp
-// CHECK-NEXT: detail/async_work_group_copy_ptr.hpp
-// CHECK-NEXT: device_event.hpp
-// CHECK-NEXT: h_item.hpp
-// CHECK-NEXT: item.hpp
 // CHECK-NEXT: kernel_handler.hpp
-// CHECK-NEXT: nd_item.hpp
-// CHECK-NEXT: nd_range.hpp
+// CHECK-NEXT: detail/kernel_desc.hpp
 // CHECK-NEXT: detail/nd_range_view.hpp
+// CHECK-NEXT: nd_range.hpp
 // CHECK-NEXT: detail/optional.hpp
 // CHECK-NEXT: detail/range_rounding.hpp
 // CHECK-NEXT: ext/oneapi/kernel_properties.hpp
@@ -134,6 +127,8 @@
 // CHECK-NEXT: ext/oneapi/properties/property.hpp
 // CHECK-NEXT: ext/oneapi/properties.hpp
 // CHECK-NEXT: ext/oneapi/properties/property_utils.hpp
+// CHECK-NEXT: item.hpp
+// CHECK-NEXT: detail/item_base.hpp
 // CHECK-NEXT: event.hpp
 // CHECK-NEXT: exception_list.hpp
 // CHECK-NEXT: ext/oneapi/device_global/device_global.hpp
@@ -167,14 +162,16 @@
 // CHECK-NEXT: ext/oneapi/work_group_scratch_memory.hpp
 // CHECK-NEXT: detail/sycl_local_mem_builtins.hpp
 // CHECK-NEXT: detail/reduction_forward.hpp
-// CHECK-NEXT: detail/ur.hpp
-// CHECK-NEXT: ur_api_funcs.def
 // CHECK-NEXT: ext/oneapi/bindless_images_interop.hpp
 // CHECK-NEXT: ext/oneapi/interop_common.hpp
 // CHECK-NEXT: ext/oneapi/bindless_images_mem_handle.hpp
 // CHECK-NEXT: ext/oneapi/experimental/free_function_traits.hpp
 // CHECK-NEXT: ext/oneapi/experimental/raw_kernel_arg.hpp
+// CHECK-NEXT: group.hpp
+// CHECK-NEXT: detail/async_work_group_copy_ptr.hpp
+// CHECK-NEXT: device_event.hpp
 // CHECK-NEXT: kernel.hpp
+// CHECK-NEXT: nd_item.hpp
 // CHECK-NEXT: sampler.hpp
 // CHECK-NEXT: sycl_span.hpp
 // CHECK-NEXT: usm/usm_pointer_info.hpp
